Refactoring + Changes from #996
