+++ map_gen/maps/space_race/market_items.lua
@@ -6,7 +6,7 @@ return {
         description = {'', 'Unlocks stone walls research'},
         sprite = 'technology/stone-walls',
         stack_limit = 1,
-        price = 250,
+        price = 500,
         disabled = true,
         disabled_reason = {'', 'DISABLED'}
     },
@@ -17,7 +17,7 @@ return {
         description = {'', 'Unlocks heavy armor research'},
         sprite = 'technology/heavy-armor',
         stack_limit = 1,
-        price = 400,
+        price = 1000,
         disabled = true,
         disabled_reason = {'', 'DISABLED'}
     },
@@ -28,7 +28,7 @@ return {
         description = {'', 'Unlocks military research'},
         sprite = 'technology/military',
         stack_limit = 1,
-        price = 100,
+        price = 1000,
         disabled = true,
         disabled_reason = {'', 'DISABLED'}
     },
@@ -39,7 +39,7 @@ return {
         description = {'', 'Unlocks military 2 research'},
         sprite = 'technology/military-2',
         stack_limit = 1,
-        price = 1000,
+        price = 10000,
         disabled = true,
         disabled_reason = {'', 'DISABLED'}
     },
@@ -50,7 +50,7 @@ return {
         description = {'', 'Unlocks military 3 research'},
         sprite = 'technology/military-3',
         stack_limit = 1,
-        price = 10000,
+        price = 25000,
         disabled = true,
         disabled_reason = {'', 'DISABLED'}
     },
@@ -61,7 +61,7 @@ return {
         description = {'', 'Unlocks military 4 research'},
         sprite = 'technology/military-3',
         stack_limit = 1,
-        price = 10000,
+        price = 50000,
         disabled = true,
         disabled_reason = {'', 'DISABLED'}
     },
@@ -72,9 +72,10 @@ return {
         description = {'', 'Tanks can only be purchased'},
         sprite = 'item/tank',
         stack_limit = 1,
-        price = 200,
+        price = 2000,
         disabled = true,
         disabled_reason = {'', 'Research tanks to unlock'}
     },
-    {price = 1, name = 'raw-fish'},
+    {price = 10, name = 'raw-fish'},
+    {price = 10, name = 'repair-pack'}
 }
